update(user): update `avatar` to `ImageSchemaDTO`. Check image inUse in user
