+++ libs/resource/src/users/dtos/index.ts
@@ -1,4 +1,3 @@
 export * from './create-user.dto';
 export * from './user-mnt-response.dto';
-export * from './address.schema.dto';
-export * from './block.schema.dto';
+export * from './user-schema.dto';
