+++ libs/resource/src/users/schemas/image.schema.ts
@@ -1,0 +1,9 @@
+import { Prop } from '@nestjs/mongoose';
+
+export class ImageSchema {
+    @Prop({ required: true, unique: true })
+    publicId: string;
+
+    @Prop({ required: true })
+    url: string;
+}
