+++ libs/resource/src/users/users.service.ts
@@ -96,4 +96,8 @@ export class UsersService {
     async isUserNameExist(userName: string): Promise<boolean> {
         return (await this.usersRepository.count({ userName })) > 0;
     }
+
+    public async isImageInUse(publicId: string): Promise<boolean> {
+        return (await this.usersRepository.count({ 'avatar.publicId': publicId })) > 0;
+    }
 }
