+++ libs/resource/src/users/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './cache.user.util';
+export * from './user.util';
